safer default
only load gcc and cmake modules in quda
better handling of directories
never do 'rm -rf *'

The following only became an issue because I used xlc by default,
and only loaded gcc/7.3.1 for quda.

For qopqdp:
something wrong with the default xl/beta-2018.10.29
I'm getting
/usr/tce/packages/xl/xl-beta-2018.10.29/xlC/16.1.1/bin/.orig/xlc_r: error: 1501-230 Internal compiler error
so we use gcc/7.3.1 for now

For qhmc:
quda.a does not have the full libstdc++ in for some reason.
So we pass additional link flags to link to libstdc++ provided by gcc-7.3.1